+++ src/steps/initializeBranchProtectionSettings.test.ts
@@ -1,9 +1,9 @@
 import { Octokit } from "octokit";
-import { SpyInstance, describe, expect, it, vi } from "vitest";
+import { MockInstance, describe, expect, it, vi } from "vitest";
 
 import { initializeBranchProtectionSettings } from "./initializeGitHubRepository/initializeBranchProtectionSettings.js";
 
-const createMockOctokit = (request: SpyInstance) =>
+const createMockOctokit = (request: MockInstance) =>
 	({
 		request,
 	}) as unknown as Octokit;