+++ src/steps/writeReadme/index.test.ts
@@ -50,23 +50,23 @@ describe("writeReadme", () => {
 			[
 			  [
 			    "README.md",
-			    "<h1 align=\\"center\\">Test Title</h1>
+			    "<h1 align="center">Test Title</h1>
 
-			<p align=\\"center\\">Test description.</p>
+			<p align="center">Test description.</p>
 
-			<p align=\\"center\\">
+			<p align="center">
 				<!-- prettier-ignore-start -->
 				<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
-				<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"All Contributors: 2\\" src=\\"https://img.shields.io/badge/all_contributors-17-21bb42.svg\\" /></a>
+				<a href="#contributors" target="_blank"><img alt="All Contributors: 2" src="https://img.shields.io/badge/all_contributors-17-21bb42.svg" /></a>
 				<!-- ALL-CONTRIBUTORS-BADGE:END -->
 				<!-- prettier-ignore-end -->
-				<a href=\\"https://codecov.io/gh/TestOwner/test-repository\\" target=\\"_blank\\"><img alt=\\"Codecov Test Coverage\\" src=\\"https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg\\"/></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md\\" target=\\"_blank\\"><img alt=\\"Contributor Covenant\\" src=\\"https://img.shields.io/badge/code_of_conduct-enforced-21bb42\\" /></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/LICENSE.md\\" target=\\"_blank\\"><img alt=\\"License: MIT\\" src=\\"https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42\\"></a>
-				<a href=\\"https://github.com/sponsors/TestFunding\\" target=\\"_blank\\"><img alt=\\"Sponsor: On GitHub\\" src=\\"https://img.shields.io/badge/sponsor-on_github-21bb42.svg\\" /></a>
-				<img alt=\\"Style: Prettier\\" src=\\"https://img.shields.io/badge/style-prettier-21bb42.svg\\" />
-				<img alt=\\"TypeScript: Strict\\" src=\\"https://img.shields.io/badge/typescript-strict-21bb42.svg\\" />
-				<img alt=\\"npm package version\\" src=\\"https://img.shields.io/npm/v/create-typescript-app?color=21bb42\\" />
+				<a href="https://codecov.io/gh/TestOwner/test-repository" target="_blank"><img alt="Codecov Test Coverage" src="https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg"/></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md" target="_blank"><img alt="Contributor Covenant" src="https://img.shields.io/badge/code_of_conduct-enforced-21bb42" /></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/LICENSE.md" target="_blank"><img alt="License: MIT" src="https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42"></a>
+				<a href="https://github.com/sponsors/TestFunding" target="_blank"><img alt="Sponsor: On GitHub" src="https://img.shields.io/badge/sponsor-on_github-21bb42.svg" /></a>
+				<img alt="Style: Prettier" src="https://img.shields.io/badge/style-prettier-21bb42.svg" />
+				<img alt="TypeScript: Strict" src="https://img.shields.io/badge/typescript-strict-21bb42.svg" />
+				<img alt="npm package version" src="https://img.shields.io/npm/v/create-typescript-app?color=21bb42" />
 			</p>
 
 			## Usage
@@ -75,9 +75,9 @@ describe("writeReadme", () => {
 			npm i test-repository
 			\`\`\`
 			\`\`\`ts
-			import { greet } from \\"test-repository\\";
+			import { greet } from "test-repository";
 
-			greet(\\"Hello, world! 💖\\");
+			greet("Hello, world! 💖");
 			\`\`\`
 
 			## Contributors
@@ -95,7 +95,7 @@ describe("writeReadme", () => {
 
 			<!-- You can remove this notice if you don't want it 🙂 no worries! -->
 
-			> 💙 This package was templated with [create-typescript-app](https://github.com/JoshuaKGoldberg/create-typescript-app).
+			> 💙 This package was templated with [\`create-typescript-app\`](https://github.com/JoshuaKGoldberg/create-typescript-app).
 			",
 			  ],
 			]
@@ -111,23 +111,23 @@ describe("writeReadme", () => {
 			[
 			  [
 			    "README.md",
-			    "<h1 align=\\"center\\">Test Title</h1>
+			    "<h1 align="center">Test Title</h1>
 
-			<p align=\\"center\\">Test description.</p>
+			<p align="center">Test description.</p>
 
-			<p align=\\"center\\">
+			<p align="center">
 				<!-- prettier-ignore-start -->
 				<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
-				<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"All Contributors: 2\\" src=\\"https://img.shields.io/badge/all_contributors-17-21bb42.svg\\" /></a>
+				<a href="#contributors" target="_blank"><img alt="All Contributors: 2" src="https://img.shields.io/badge/all_contributors-17-21bb42.svg" /></a>
 				<!-- ALL-CONTRIBUTORS-BADGE:END -->
 				<!-- prettier-ignore-end -->
-				<a href=\\"https://codecov.io/gh/TestOwner/test-repository\\" target=\\"_blank\\"><img alt=\\"Codecov Test Coverage\\" src=\\"https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg\\"/></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md\\" target=\\"_blank\\"><img alt=\\"Contributor Covenant\\" src=\\"https://img.shields.io/badge/code_of_conduct-enforced-21bb42\\" /></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/LICENSE.md\\" target=\\"_blank\\"><img alt=\\"License: MIT\\" src=\\"https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42\\"></a>
-				<a href=\\"https://github.com/sponsors/TestFunding\\" target=\\"_blank\\"><img alt=\\"Sponsor: On GitHub\\" src=\\"https://img.shields.io/badge/sponsor-on_github-21bb42.svg\\" /></a>
-				<img alt=\\"Style: Prettier\\" src=\\"https://img.shields.io/badge/style-prettier-21bb42.svg\\" />
-				<img alt=\\"TypeScript: Strict\\" src=\\"https://img.shields.io/badge/typescript-strict-21bb42.svg\\" />
-				<img alt=\\"npm package version\\" src=\\"https://img.shields.io/npm/v/create-typescript-app?color=21bb42\\" />
+				<a href="https://codecov.io/gh/TestOwner/test-repository" target="_blank"><img alt="Codecov Test Coverage" src="https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg"/></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md" target="_blank"><img alt="Contributor Covenant" src="https://img.shields.io/badge/code_of_conduct-enforced-21bb42" /></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/LICENSE.md" target="_blank"><img alt="License: MIT" src="https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42"></a>
+				<a href="https://github.com/sponsors/TestFunding" target="_blank"><img alt="Sponsor: On GitHub" src="https://img.shields.io/badge/sponsor-on_github-21bb42.svg" /></a>
+				<img alt="Style: Prettier" src="https://img.shields.io/badge/style-prettier-21bb42.svg" />
+				<img alt="TypeScript: Strict" src="https://img.shields.io/badge/typescript-strict-21bb42.svg" />
+				<img alt="npm package version" src="https://img.shields.io/npm/v/create-typescript-app?color=21bb42" />
 			</p>
 
 			## Usage
@@ -136,9 +136,9 @@ describe("writeReadme", () => {
 			npm i test-repository
 			\`\`\`
 			\`\`\`ts
-			import { greet } from \\"test-repository\\";
+			import { greet } from "test-repository";
 
-			greet(\\"Hello, world! 💖\\");
+			greet("Hello, world! 💖");
 			\`\`\`
 
 			## Contributors
@@ -156,7 +156,7 @@ describe("writeReadme", () => {
 
 			<!-- You can remove this notice if you don't want it 🙂 no worries! -->
 
-			> 💙 This package was templated with [create-typescript-app](https://github.com/JoshuaKGoldberg/create-typescript-app).
+			> 💙 This package was templated with [\`create-typescript-app\`](https://github.com/JoshuaKGoldberg/create-typescript-app).
 			",
 			  ],
 			]
@@ -175,23 +175,23 @@ describe("writeReadme", () => {
 			[
 			  [
 			    "README.md",
-			    "<h1 align=\\"center\\">Test Title</h1>
+			    "<h1 align="center">Test Title</h1>
 
-			<p align=\\"center\\">Test description.</p>
+			<p align="center">Test description.</p>
 
-			<p align=\\"center\\">
+			<p align="center">
 				<!-- prettier-ignore-start -->
 				<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
-				<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"All Contributors: 2\\" src=\\"https://img.shields.io/badge/all_contributors-17-21bb42.svg\\" /></a>
+				<a href="#contributors" target="_blank"><img alt="All Contributors: 2" src="https://img.shields.io/badge/all_contributors-17-21bb42.svg" /></a>
 				<!-- ALL-CONTRIBUTORS-BADGE:END -->
 				<!-- prettier-ignore-end -->
-				<a href=\\"https://codecov.io/gh/TestOwner/test-repository\\" target=\\"_blank\\"><img alt=\\"Codecov Test Coverage\\" src=\\"https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg\\"/></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md\\" target=\\"_blank\\"><img alt=\\"Contributor Covenant\\" src=\\"https://img.shields.io/badge/code_of_conduct-enforced-21bb42\\" /></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/LICENSE.md\\" target=\\"_blank\\"><img alt=\\"License: MIT\\" src=\\"https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42\\"></a>
-				<a href=\\"https://github.com/sponsors/TestFunding\\" target=\\"_blank\\"><img alt=\\"Sponsor: On GitHub\\" src=\\"https://img.shields.io/badge/sponsor-on_github-21bb42.svg\\" /></a>
-				<img alt=\\"Style: Prettier\\" src=\\"https://img.shields.io/badge/style-prettier-21bb42.svg\\" />
-				<img alt=\\"TypeScript: Strict\\" src=\\"https://img.shields.io/badge/typescript-strict-21bb42.svg\\" />
-				<img alt=\\"npm package version\\" src=\\"https://img.shields.io/npm/v/create-typescript-app?color=21bb42\\" />
+				<a href="https://codecov.io/gh/TestOwner/test-repository" target="_blank"><img alt="Codecov Test Coverage" src="https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg"/></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md" target="_blank"><img alt="Contributor Covenant" src="https://img.shields.io/badge/code_of_conduct-enforced-21bb42" /></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/LICENSE.md" target="_blank"><img alt="License: MIT" src="https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42"></a>
+				<a href="https://github.com/sponsors/TestFunding" target="_blank"><img alt="Sponsor: On GitHub" src="https://img.shields.io/badge/sponsor-on_github-21bb42.svg" /></a>
+				<img alt="Style: Prettier" src="https://img.shields.io/badge/style-prettier-21bb42.svg" />
+				<img alt="TypeScript: Strict" src="https://img.shields.io/badge/typescript-strict-21bb42.svg" />
+				<img alt="npm package version" src="https://img.shields.io/npm/v/create-typescript-app?color=21bb42" />
 			</p>
 
 			## Usage
@@ -200,9 +200,9 @@ describe("writeReadme", () => {
 			npm i test-repository
 			\`\`\`
 			\`\`\`ts
-			import { greet } from \\"test-repository\\";
+			import { greet } from "test-repository";
 
-			greet(\\"Hello, world! 💖\\");
+			greet("Hello, world! 💖");
 			\`\`\`
 
 			## Contributors
@@ -220,7 +220,7 @@ describe("writeReadme", () => {
 
 			<!-- You can remove this notice if you don't want it 🙂 no worries! -->
 
-			> 💙 This package was templated with [create-typescript-app](https://github.com/JoshuaKGoldberg/create-typescript-app).
+			> 💙 This package was templated with [\`create-typescript-app\`](https://github.com/JoshuaKGoldberg/create-typescript-app).
 			",
 			  ],
 			]
@@ -273,24 +273,24 @@ describe("writeReadme", () => {
 			[
 			  [
 			    "README.md",
-			    "<h1 align=\\"center\\">Test Title</h1>
+			    "<h1 align="center">Test Title</h1>
 
-			<p align=\\"center\\">Test description.</p>
+			<p align="center">Test description.</p>
 
-			<p align=\\"center\\">
+			<p align="center">
 				<!-- prettier-ignore-start -->
 				<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
-				<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"All Contributors: 2\\" src=\\"https://img.shields.io/badge/all_contributors-17-21bb42.svg\\" /></a>
+				<a href="#contributors" target="_blank"><img alt="All Contributors: 2" src="https://img.shields.io/badge/all_contributors-17-21bb42.svg" /></a>
 				<!-- ALL-CONTRIBUTORS-BADGE:END -->
 				<!-- prettier-ignore-end -->
-				<a href=\\"https://codecov.io/gh/TestOwner/test-repository\\" target=\\"_blank\\"><img alt=\\"Codecov Test Coverage\\" src=\\"https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg\\"/></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md\\" target=\\"_blank\\"><img alt=\\"Contributor Covenant\\" src=\\"https://img.shields.io/badge/code_of_conduct-enforced-21bb42\\" /></a>
-				<a href=\\"https://github.com/TestOwner/test-repository/blob/main/LICENSE.md\\" target=\\"_blank\\"><img alt=\\"License: MIT\\" src=\\"https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42\\"></a>
-				<a href=\\"https://github.com/sponsors/TestFunding\\" target=\\"_blank\\"><img alt=\\"Sponsor: On GitHub\\" src=\\"https://img.shields.io/badge/sponsor-on_github-21bb42.svg\\" /></a>
-				<img alt=\\"Style: Prettier\\" src=\\"https://img.shields.io/badge/style-prettier-21bb42.svg\\" />
-				<img alt=\\"TypeScript: Strict\\" src=\\"https://img.shields.io/badge/typescript-strict-21bb42.svg\\" />
-				<img alt=\\"npm package version\\" src=\\"https://img.shields.io/npm/v/create-typescript-app?color=21bb42\\" />
-				<a href=\\"#contributors\\" target=\\"_blank\\"><img alt=\\"All Contributors: 2\\" src=\\"https://img.shields.io/badge/all_contributors-17-21bb42.svg\\" /></a>
+				<a href="https://codecov.io/gh/TestOwner/test-repository" target="_blank"><img alt="Codecov Test Coverage" src="https://codecov.io/gh/TestOwner/test-repository/branch/main/graph/badge.svg"/></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/.github/CODE_OF_CONDUCT.md" target="_blank"><img alt="Contributor Covenant" src="https://img.shields.io/badge/code_of_conduct-enforced-21bb42" /></a>
+				<a href="https://github.com/TestOwner/test-repository/blob/main/LICENSE.md" target="_blank"><img alt="License: MIT" src="https://img.shields.io/github/license/TestOwner/test-repository?color=21bb42"></a>
+				<a href="https://github.com/sponsors/TestFunding" target="_blank"><img alt="Sponsor: On GitHub" src="https://img.shields.io/badge/sponsor-on_github-21bb42.svg" /></a>
+				<img alt="Style: Prettier" src="https://img.shields.io/badge/style-prettier-21bb42.svg" />
+				<img alt="TypeScript: Strict" src="https://img.shields.io/badge/typescript-strict-21bb42.svg" />
+				<img alt="npm package version" src="https://img.shields.io/npm/v/create-typescript-app?color=21bb42" />
+				<a href="#contributors" target="_blank"><img alt="All Contributors: 2" src="https://img.shields.io/badge/all_contributors-17-21bb42.svg" /></a>
 			</p>
 
 			## Usage
@@ -299,9 +299,9 @@ describe("writeReadme", () => {
 			npm i test-repository
 			\`\`\`
 			\`\`\`ts
-			import { greet } from \\"test-repository\\";
+			import { greet } from "test-repository";
 
-			greet(\\"Hello, world! 💖\\");
+			greet("Hello, world! 💖");
 			\`\`\`
 
 			## Contributors